git-automation/requirements.txt
beautifulsoup4==4.12.3
gitdb==4.0.11
GitPython==3.1.42
smmap==5.0.1
soupsieve==2.5
